dissect-cf-core/src/main/java/hu/vio/thesis/layers/predictor/requirements.txt
fbprophet==0.7.1
keras==2.11.0
matplotlib==3.7.1
neuralprophet==0.5.1
numpy==1.24.3
pandas==1.2.5
plotly==5.13.0
prophet==1.1.2
pyaf==4.0
pywebview==4.0.2
scikit_learn==1.1.3
scipy==1.10.1
skforecast==0.6.0
sktime==0.18.0
statsmodels==0.13.5
tensorflow==2.11.0
tensorflow_intel==2.11.0
thesis==0.0.01
xgboost==1.7.5
